requirements.txt
Flask>=2.0
requests>=2.28
python-dotenv>=0.21
yara-python>=4.0.0
python-magic>=0.4.27
